refactor: Enhance API configuration and simplify codebase
- Refactored `configureDatabase` to clarify responsibilities and improve DI bindings.
- Updated `configureGraphQL` to simplify schema and resolver setup using custom hooks.
- Enhanced CORS policy setup in `configureCors` with more concise constants and comments.
- Streamlined HTTP response compression and caching in `configureCompression` and `configureCaching`.
- Simplified logging configuration in `configureCallLogging` and improved readability.
- Modularized registration of GraphQL models, resolvers, and hooks.
- Improved readiness and modularity of MCP server configuration (`configureMcp`) by introducing tool extensions and externalized server properties.
- Consolidated imports across multiple files for better clarity and reduced clutter in test and application code.
- Made consistent usage of domain-level extensions like `toGraphQlProduct` and `toMcpProduct`.
