+++ server/src/main/kotlin/dev/aoriani/ecomm/config/cache.kt
@@ -4,13 +4,17 @@ import io.ktor.server.application.Application
 import io.ktor.server.application.install
 import io.ktor.server.plugins.conditionalheaders.ConditionalHeaders
 
+
 /**
- * Configures caching headers for the application.
+ * Configure HTTP conditional requests (ETag/Last-Modified).
  *
- * This method installs:
- * - The `ConditionalHeaders` feature to enable browser caching through ETags and Last-Modified headers
+ * Installs the Ktor `ConditionalHeaders` plugin which:
+ * - Adds ETag and Last-Modified support where applicable
+ * - Responds with 304 Not Modified when the resource hasn't changed
+ * - Reduces bandwidth and speeds up repeat requests
  *
- * Conditional headers are enabled to improve caching efficiency and reduce unnecessary transfers.
+ * Note: This does not add Cache-Control headers. Use the `CachingHeaders`
+ * plugin if you need explicit Cache-Control directives.
  */
 internal fun Application.configureCaching() {
     install(ConditionalHeaders)
