+++ server/src/main/kotlin/dev/aoriani/ecomm/config/graphql.kt
@@ -10,17 +10,16 @@ import io.ktor.server.application.install
 import io.ktor.server.plugins.di.dependencies
 
 /**
- * Configures the GraphQL server for the application.
+ * Configure the GraphQL server.
  *
- * This method sets up the `GraphQL` feature with the following:
- * - Schema generation using specified package paths.
- * - Query resolvers to handle product-related queries.
- * - Custom hooks for additional schema customization or behavior adjustments.
+ * Installs Expedia Group's Ktor `GraphQL` plugin and configures:
+ * - Schema packages: `dev.aoriani.ecomm.presentation.graphql.models`, `java.math`.
+ * - Query resolvers: `ProductQuery` wired with domain use cases.
+ * - Schema hooks: `ProductSchemaGeneratorHooks` for customizations.
  *
- * The schema is dynamically built by scanning classes within the provided packages for GraphQL types
- * and directives, and associating them with the specified query resolvers.
- *
- * Dependencies required for query resolver initialization are retrieved using the application's dependency injection mechanism.
+ * The schema is generated by scanning the configured packages for GraphQL
+ * types/directives and binding them to the provided resolvers. Use cases are
+ * resolved via Ktor's DI (`dependencies`).
  */
 internal fun Application.configureGraphQL() {
     install(GraphQL) {
