+++ server/src/main/kotlin/dev/aoriani/ecomm/presentation/mcp/tools/GetAllProductsTool.kt
@@ -1,14 +1,14 @@
 package dev.aoriani.ecomm.presentation.mcp.tools
 
 import dev.aoriani.ecomm.domain.usecases.GetAllProductsUseCase
+import dev.aoriani.ecomm.domain.usecases.invoke
+import dev.aoriani.ecomm.presentation.mcp.models.Products
+import dev.aoriani.ecomm.presentation.mcp.models.toMcpProduct
 import io.modelcontextprotocol.kotlin.sdk.CallToolRequest
 import io.modelcontextprotocol.kotlin.sdk.CallToolResult
 import io.modelcontextprotocol.kotlin.sdk.TextContent
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.encodeToJsonElement
-import dev.aoriani.ecomm.domain.usecases.invoke
-import dev.aoriani.ecomm.presentation.mcp.models.Products
-import dev.aoriani.ecomm.presentation.mcp.models.toMcpProduct
 import kotlinx.serialization.json.jsonObject
 import kotlin.reflect.KClass
