+++ server/src/test/kotlin/dev/aoriani/ecomm/GraphQlEndpointTest.kt
@@ -1,28 +1,17 @@
 package dev.aoriani.ecomm
 
-import com.expediagroup.graphql.generator.scalars.ID
 import dev.aoriani.ecomm.domain.models.Product
 import dev.aoriani.ecomm.domain.repositories.ProductRepository
-import io.ktor.client.request.post
-import io.ktor.client.request.setBody
-import io.ktor.client.statement.bodyAsText
-import io.ktor.http.ContentType
-import io.ktor.http.HttpStatusCode
-import io.ktor.http.contentType
-import io.ktor.http.withCharset
-import io.ktor.server.config.MapApplicationConfig
-import io.ktor.server.plugins.di.dependencies
-import io.ktor.server.testing.ApplicationTestBuilder
-import io.ktor.server.testing.testApplication
-import io.ktor.utils.io.charsets.Charsets
+import io.ktor.client.request.*
+import io.ktor.client.statement.*
+import io.ktor.http.*
+import io.ktor.server.config.*
+import io.ktor.server.plugins.di.*
+import io.ktor.server.testing.*
+import io.ktor.utils.io.charsets.*
 import io.mockk.coEvery
 import io.mockk.mockk
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.buildJsonObject
-import kotlinx.serialization.json.jsonArray
-import kotlinx.serialization.json.jsonObject
-import kotlinx.serialization.json.jsonPrimitive
+import kotlinx.serialization.json.*
 import java.math.BigDecimal
 import kotlin.test.Test
 import kotlin.test.assertEquals
